backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0

# Database
motor==3.3.2  # MongoDB async driver
pymongo==4.6.0

# NLP Libraries
google-generativeai==0.8.3
groq==0.11.0
httpx==0.27.0
nltk==3.8.1
scikit-learn==1.3.2

# Document Processing
PyMuPDF==1.23.8  # PDF processing
python-docx==1.1.0  # DOCX processing
python-pptx==0.6.23  # PPTX processing
pytesseract==0.3.10  # OCR
Pillow==10.1.0  # Image processing

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.1
aiofiles==23.2.1
